end-to-end tests passing

+++ solidity/contracts/Mailbox.sol
@@ -174,8 +174,9 @@ contract Mailbox is
     /**
      * @notice Returns the number of inserted leaves in the tree
      */
-    function count() public view returns (uint256) {
-        return tree.count;
+    function count() public view returns (uint32) {
+        // count cannot exceed 2**TREE_DEPTH, see MerkleLib.sol
+        return uint32(tree.count);
     }
 
     /**
